Use 'clap' crate for argument parsing
