+++ Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clap = { version = "4.5.30", features = ["derive"] }
 rand = "0.8.5"
 serde = "1.0.197"
 serde_json = "1.0.115"
